X64: Optimize stubs for incremental marking.
R=erik.corry@gmail.com
BUG=v8:1457

Review URL: http://codereview.chromium.org/7676008

git-svn-id: http://v8.googlecode.com/svn/branches/experimental@8970 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ gc/src/incremental-marking.h
@@ -180,7 +180,7 @@ class IncrementalMarking {
       if (allocation_marking_factor_ < kFastMarking) {
         if (FLAG_trace_gc) {
           PrintF("Increasing marking speed to %d due to high promotion rate\n",
-                 kFastMarking);
+                 static_cast<int>(kFastMarking));
         }
         allocation_marking_factor_ = kFastMarking;
       }
